ai_agent/requirements.txt
# Core LangChain
langchain==0.3.13
langchain-core==0.3.29
langchain-community==0.3.13

# LangChain Integrations
langchain-openai==0.2.14

# Vector Stores
faiss-cpu==1.9.0

# Basic Tools
requests==2.32.3
python-dotenv==1.0.1
tiktoken==0.8.0

# OpenAI
openai==1.58.1

# Basic utilities
numpy==1.26.4
pydantic==2.10.4
tenacity==9.0.0 